Move stream utilities to a separate folder.
