+++ modules/module-mongodb-storage/src/storage/implementation/MongoWriteCheckpointAPI.ts
@@ -1,12 +1,9 @@
 import { mongo } from '@powersync/lib-service-mongodb';
 import * as framework from '@powersync/lib-services-framework';
 import { storage, WatchUserWriteCheckpointOptions, WriteCheckpointResult } from '@powersync/service-core';
-import { AbortError } from 'ix/aborterror.js';
-import { AsyncSink } from 'ix/asynciterable/asynciterablex.js';
 import { PowerSyncMongo } from './db.js';
 import { CustomWriteCheckpointDocument, WriteCheckpointDocument } from './models.js';
-
-import { wrapWithAbort } from 'ix/asynciterable/operators/withabort.js';
+import { Demultiplexer, DemultiplexerValue } from '@powersync/service-core/src/streams/Demultiplexer.js';
 
 export type MongoCheckpointAPIOptions = {
   db: PowerSyncMongo;
@@ -110,7 +107,7 @@ export class MongoWriteCheckpointAPI implements storage.WriteCheckpointAPI {
     }
   }
 
-  private sharedIter = new FilteredIterable<WriteCheckpointResult>((signal) => {
+  private sharedIter = new Demultiplexer<WriteCheckpointResult>((signal) => {
     const clusterTimePromise = (async () => {
       const hello = await this.db.db.command({ hello: 1 });
       // Note: This is not valid on sharded clusters.
@@ -180,7 +177,7 @@ export class MongoWriteCheckpointAPI implements storage.WriteCheckpointAPI {
   private async *watchAllManagedWriteCheckpoints(
     clusterTimePromise: Promise<mongo.BSON.Timestamp>,
     signal: AbortSignal
-  ): AsyncGenerator<FilterIterableValue<WriteCheckpointResult>> {
+  ): AsyncGenerator<DemultiplexerValue<WriteCheckpointResult>> {
     const clusterTime = await clusterTimePromise;
 
     const stream = this.db.write_checkpoints.watch(
@@ -364,135 +361,4 @@ export async function batchCreateCustomWriteCheckpoints(
     })),
     {}
   );
-}
-
-export interface FilterIterableValue<T> {
-  key: string;
-  value: T;
-}
-
-interface FilteredIterableSource<T> {
-  iterator: AsyncIterable<FilterIterableValue<T>>;
-  getFirstValue(key: string): Promise<T>;
-}
-
-type FilteredIterableSourceFactory<T> = (signal: AbortSignal) => FilteredIterableSource<T>;
-
-export class FilteredIterable<T> {
-  private subscribers: Map<string, Set<AsyncSink<T>>> | undefined = undefined;
-  private abortController: AbortController | undefined = undefined;
-  private currentSource: FilteredIterableSource<T> | undefined = undefined;
-
-  constructor(private source: FilteredIterableSourceFactory<T>) {}
-
-  private start(filter: string, sink: AsyncSink<T>) {
-    const abortController = new AbortController();
-    const listeners = new Map();
-    listeners.set(filter, new Set([sink]));
-
-    this.abortController = abortController;
-    this.subscribers = listeners;
-
-    const source = this.source(abortController.signal);
-    this.currentSource = source;
-    this.loop(source, abortController, listeners);
-    return source;
-  }
-
-  private async loop(
-    source: FilteredIterableSource<T>,
-    abortController: AbortController,
-    sinks: Map<string, Set<AsyncSink<T>>>
-  ) {
-    try {
-      for await (let doc of source.iterator) {
-        if (abortController.signal.aborted || sinks.size == 0) {
-          throw new AbortError();
-        }
-        const key = doc.key;
-        const keySinks = sinks.get(key);
-        if (keySinks == null) {
-          continue;
-        }
-
-        for (let sink of keySinks) {
-          sink.write(doc.value);
-        }
-      }
-
-      // End of stream
-      for (let keySinks of sinks.values()) {
-        for (let sink of keySinks) {
-          sink.end();
-        }
-      }
-    } catch (e) {
-      // Just in case the error is not from the source
-      abortController.abort();
-
-      for (let keySinks of sinks.values()) {
-        for (let sink of keySinks) {
-          sink.error(e);
-        }
-      }
-    } finally {
-      // Clear state, so that a new subscription may be started
-      if (this.subscribers === sinks) {
-        this.subscribers = undefined;
-        this.abortController = undefined;
-        this.currentSource = undefined;
-      }
-    }
-  }
-
-  private removeSink(key: string, sink: AsyncSink<T>) {
-    const existing = this.subscribers?.get(key);
-    if (existing == null) {
-      return;
-    }
-    existing.delete(sink);
-    if (existing.size == 0) {
-      this.subscribers!.delete(key);
-    }
-
-    if (this.subscribers?.size == 0) {
-      // This is not immediate - there may be a delay until it is fully stopped,
-      // depending on the underlying source.
-      this.abortController?.abort();
-      this.subscribers = undefined;
-      this.abortController = undefined;
-      this.currentSource = undefined;
-    }
-  }
-
-  private addSink(key: string, sink: AsyncSink<T>) {
-    if (this.currentSource == null) {
-      return this.start(key, sink);
-    } else {
-      const existing = this.subscribers!.get(key);
-      if (existing != null) {
-        existing.add(sink);
-      } else {
-        this.subscribers!.set(key, new Set([sink]));
-      }
-      return this.currentSource;
-    }
-  }
-
-  async *subscribe(key: string, signal: AbortSignal): AsyncIterable<T> {
-    const sink = new AsyncSink<T>();
-    // Important that we register the sink before calling getFirstValue().
-    const source = this.addSink(key, sink);
-    try {
-      const firstValue = await source.getFirstValue(key);
-      yield firstValue;
-      yield* wrapWithAbort(sink, signal);
-    } finally {
-      this.removeSink(key, sink);
-    }
-  }
-
-  get active() {
-    return this.subscribers != null;
-  }
 }
