+++ packages/service-core/src/index.ts
@@ -38,3 +38,6 @@ export * as system from './system/system-index.js';
 
 export * from './util/util-index.js';
 export * as utils from './util/util-index.js';
+
+export * from './streams/streams-index.js';
+export * as streams from './streams/streams-index.js';
