+++ packages/service-core/src/streams/LastValueSink.ts
