+++ packages/service-core/src/streams/streams-index.ts
@@ -1,0 +1,4 @@
+export * from './merge.js';
+export * from './Demultiplexer.js';
+export * from './LastValueSink.js';
+export * from './BroadcastIterable.js';
