+++ packages/service-core/src/sync/sync-index.ts
@@ -1,6 +1,3 @@
-export * from './BroadcastIterable.js';
-export * from './LastValueSink.js';
-export * from './merge.js';
 export * from './RequestTracker.js';
 export * from './safeRace.js';
 export * from './sync.js';
