+++ packages/service-core/src/sync/sync.ts
@@ -9,7 +9,7 @@ import * as util from '../util/util-index.js';
 
 import { logger } from '@powersync/lib-services-framework';
 import { BucketChecksumState } from './BucketChecksumState.js';
-import { mergeAsyncIterables } from './merge.js';
+import { mergeAsyncIterables } from '../streams/streams-index.js';
 import { acquireSemaphoreAbortable, settledPromise, tokenStream, TokenStreamOptions } from './util.js';
 import { SyncContext } from './SyncContext.js';
 import { RequestTracker } from './RequestTracker.js';
